docs: capture the LongArray-storage decision with verbatim Kotlin source citations (#11)
* docs: add why-longarray-storage.md and wire it into existing references

Background: the recent benchmark migration to kotlinx-benchmark surfaced
that `loadWordSweepUnsigned` (using stdlib `UByteArray`) ran ~14% slower
than `loadWordSweepSigned` (using `ByteArray + and 0xFF`) on Kotlin/Native
macOS arm64. Adding a third `loadWordSweepLongArray` variant produced:

    loadWordSweepLongArray   2.172 µs/op  ← klang's PackedBuffer storage
    loadWordSweepSigned      4.274 µs/op  ← naive ByteArray + mask
    loadWordSweepUnsigned    4.846 µs/op  ← stdlib UByteArray

LongArray is 1.97x faster than ByteArray+mask and 2.23x faster than
UByteArray. CIs do not overlap.

The reason is structural: `Byte` and `Short` have no `shl`/`shr`/`ushr`
operators in stdlib (verified in v2.3.21
`libraries/stdlib/src/kotlin/Primitives.kt` — only Int and Long have
shifts; the experimental bitwise ops on Byte at
`libraries/stdlib/src/kotlin/experimental/bitwiseOperations.kt` cover
and/or/xor/inv only and internally promote to Int via the
sign-extending `Byte.toInt()`). `UByteArray.get` is non-`inline`
because it implements `Collection<UByte>`, so each read pays a
function-call boundary; `UByte.toInt()` is literally
`data.toInt() and 0xFF`. So the "unsigned" path does the same masking
work as the "signed" path *plus* a function-call boundary. The
`LongArray` path avoids `Byte` entirely — `Long.ushr` returns `Long`
with no type promotion, and a single `Long` carries 8 bytes so a
4-byte load is one slot read + shift + mask.

Klang already chose LongArray storage (PackedBuffer / GlobalHeap on
both Native and JS). This commit captures the *why* — with v2.3.21
stdlib line-number citations and the benchmark numbers — so future
readers don't have to re-do the investigation, and so anyone tempted
to "modernise" to UByteArray has a number to argue against.

## docs/general/why-longarray-storage.md (new, ≈ 350 lines)

The flagship document. Sections:

  - TL;DR with the three structural reasons
  - Why we looked (the benchmark anomaly that started it)
  - Where the information lives in the Kotlin v2.3.21 source
    (file paths, line numbers, grep commands the next reader can run)
  - What it means at the storage level (3-way comparison: ByteArray,
    UByteArray, LongArray, with operation counts per word load)
  - Empirical: three storage shapes, same work (the benchmark table)
  - Why this is the hard floor, not a peephole optimisation
    (stdlib API gap + inline-class boundary + Native codegen quality)
  - What klang does as a result
    (PackedBuffer, GlobalHeap actuals, byte-free hot math layer,
    legitimate ByteArray boundaries for user interop)
  - How to verify these claims yourself (mirror path + grep recipes
    + benchmark rerun command)
  - Mark Adler / zlib / 16-bit C — why this isn't pedantry, it's
    structural defence
  - Cross-references

## docs/general/philosophy-and-design.md

The "Forbidden practices" section already prohibited raw bitwise
operators outside BitShiftEngine and hard-coded masks. Added a
paragraph explaining *why* those rules exist (no shifts on Byte/Short
in stdlib + sign-extending Byte.toInt) with a link to the new doc for
the full investigation.

## src/commonMain/.../PackedBuffer.kt

Expanded the header KDoc from a brief "Why LongArray?" rationale into
a structured pointer at the new doc, with the line-number citations
called out in the comment so anyone reading the source can navigate
straight to the proof in stdlib.

## src/{nativeMain,jsMain}/.../GlobalHeap{Native,Js}.kt

Reformatted the existing one-liner rationale comment to point at the
new doc and call out the empirical 2x speedup. Same intent as before,
just made explicit.

## Verification

  ./gradlew compileKotlinMacosArm64 macosArm64Test  -> green

* audit: remove 7 redundant type-conversion round-trips + add audit document

A project-wide audit of unnecessary type conversions, run downstream of
the LongArray-substrate decision. Two outputs: tier-1 simplifications
applied directly, tier-2 design questions documented in
docs/general/type-conversion-audit.md.

## Tier-1 fixes (this commit)

Pattern 1: `Long.toInt().toShort()` -> `Long.toShort()`
  Same bit pattern, one fewer narrowing op. Long.toShort() is defined
  to return the low 16 bits — exactly what the chain produces.

  src/commonMain/kotlin/io/github/kotlinmania/klang/bitwise/BitShiftEngine.kt:1410, 1436, 1462, 1500
  src/commonMain/kotlin/io/github/kotlinmania/klang/mem/PackedBuffer.kt:102

Pattern 2: `Short.toInt().toLong()` -> `Short.toLong()`
  Short.toLong() is the direct sign-extending widening. The middle
  Int step did sign-extend-to-32 followed by sign-extend-to-64, two
  ops where one suffices.

  src/commonMain/kotlin/io/github/kotlinmania/klang/internal/runtime/Runtime.kt:125, 126
  src/commonMain/kotlin/io/github/kotlinmania/klang/fp/CBF16.kt:62, 68

7 conversion ops removed. Identical semantics. Tests green
(./gradlew macosArm64Test -> 439/0/0).

## Tier-2 documented (no code change yet)

Two API-shape questions captured in
docs/general/type-conversion-audit.md as proposals for separate PRs:

  A. PackedBuffer.setShort(addr, value: Short) probably should take
     Int to match setByte(addr, value: Int). Currently every 16-bit
     store does Int -> Short -> Int through-and-back via the type
     system. Affects ~20 call sites in F16Math, Float128Math,
     BitShiftEngine.

  B. BitShiftEngine could expose Int-overload shifts (leftShift32,
     rightShift32, unsignedRightShift32) so 32-bit operands don't
     pay Int->Long widen and Long->Int narrow on every call. Affects
     ~30 call sites in CBF16, HexShift, BitwiseOps, BitPrimitives,
     Float32Math.

Both deferred for benchmark-backed evaluation.

## Tier-3 documented (intentional patterns, not bugs)

Three patterns captured to forestall future 'cleanup' that would
remove them by mistake:

  - .toLong() at engine call sites (the engine API requires Long)
  - (bits and 0xFF).toUByte() in CE8M0/CUE4M3 (documents the
    byte-width assumption)
  - HeapUInt128.toIntArray() in Float128Math (intentional snapshot
    for read-modify-write)

## docs/general/type-conversion-audit.md (new)

Structured report with the three tiers, file-and-line table for tier-1
fixes, expanded rationale for tier-2 proposals, the grep recipes used
to find these patterns (so the audit is reproducible), and
cross-references to why-longarray-storage.md and philosophy-and-design.md.

## Verification

  ./gradlew compileKotlinMacosArm64 macosArm64Test  -> green (439/0/0)

* audit: append empirical benchmark results to type-conversion audit

The TypeConversionBenchmark was added in PR #10 (benchmark branch)
covering all three audit tiers. Results:

  Tier 1 (round-trips): ~0–2% within-noise — the audit's simplifications
  are code-clarity wins, the compiler peephole-optimises away the
  redundant intermediate Int on Kotlin/Native macOS arm64. No regression
  risk, no measurable speed change.

  Tier 2-A (setShort API shape): ~7% faster with Int parameter, CIs
  disjoint. Modest but repeatable; worth landing.

  Tier 2-B (engine Long-boundary): 40–170x slower for engine.leftShift /
  engine.unsignedRightShift vs direct Kotlin shifts. The per-call
  overhead (ShiftResult allocation, mode branch, Long widen/narrow)
  dwarfs the actual shift work. This promotes the Int-overload
  proposal from 'design question' to 'serious performance issue.'

Both Tier-2 items now have empirical justification and should land
as their own PRs after this audit doc merges.

+++ src/commonMain/kotlin/io/github/kotlinmania/klang/bitwise/BitShiftEngine.kt
@@ -1407,7 +1407,7 @@ class BitShiftEngine(
         val result = leftShift(value, bits)
         when (bitWidth) {
             8 -> buffer.setByte(addr, result.value.toInt())
-            16 -> buffer.setShort(addr, result.value.toInt().toShort())
+            16 -> buffer.setShort(addr, result.value.toShort())
             32 -> buffer.setInt(addr, result.value.toInt())
             64 -> buffer.setLong(addr, result.value)
         }
@@ -1433,7 +1433,7 @@ class BitShiftEngine(
         val result = rightShift(value, bits)
         when (bitWidth) {
             8 -> buffer.setByte(addr, result.value.toInt())
-            16 -> buffer.setShort(addr, result.value.toInt().toShort())
+            16 -> buffer.setShort(addr, result.value.toShort())
             32 -> buffer.setInt(addr, result.value.toInt())
             64 -> buffer.setLong(addr, result.value)
         }
@@ -1459,7 +1459,7 @@ class BitShiftEngine(
         val result = unsignedRightShift(value, bits)
         when (bitWidth) {
             8 -> buffer.setByte(addr, result.value.toInt())
-            16 -> buffer.setShort(addr, result.value.toInt().toShort())
+            16 -> buffer.setShort(addr, result.value.toShort())
             32 -> buffer.setInt(addr, result.value.toInt())
             64 -> buffer.setLong(addr, result.value)
         }
@@ -1497,7 +1497,7 @@ class BitShiftEngine(
     fun store(buffer: PackedBuffer, addr: Int, value: Long) {
         when (bitWidth) {
             8 -> buffer.setByte(addr, (value and 0xFF).toInt())
-            16 -> buffer.setShort(addr, (value and 0xFFFF).toInt().toShort())
+            16 -> buffer.setShort(addr, (value and 0xFFFF).toShort())
             32 -> buffer.setInt(addr, (value and 0xFFFFFFFFL).toInt())
             64 -> buffer.setLong(addr, value)
         }
